Avoid hardcoding when creating files
